requirements.txt
Pillow==9.4.0
vtf2img==0.1.0